refactored a file name and made it to where a user can stream from twitch or mixer
